requirements.txt
APScheduler==3.6.3
asgiref==3.3.4
bitcoin==1.1.42
blockcypher==1.0.91
certifi==2020.12.5
chardet==4.0.0
Django==3.2.2
django-cryptocurrency-payment==0.1.1
forex-python==1.5
idna==2.10
merchant-wallet==0.1.0
python-dateutil==2.8.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
